feat: split bottom sheet content into scenario components
